fix tests and remove logs
